requirements.txt
streamlit
faiss-cpu
openai
sentence-transformers
pypdf
beautifulsoup4
requests
numpy
pickle-mixin
